safety loop: increase to 5
False positives cannot happen (if the device is not on the network we
will not find it), on the other hand, in practice it seems the device
can easily be on the network on not be found. Since checking 3 times
seems to be barely acceptable, and we're not in a rush to detect, let's
increase this to 5.